refactor(react-app): change the App function into arrow function, and change the import of App function
